refactored Bar and Drink to match bars in database
