+++ src/data.json
@@ -2,7 +2,7 @@
 "drinks": {
     "-NpHWPZX3UPujdq_bWkx": {
         "barName": "Stampede Cocktail Club",
-        "barID": "1ea8beea-6dd9-48ea-b279-f40efd36fdf6",
+        "barID": "-Nr7lOhIhP11igBsjqOp",
         "drinkName": "85% Euphoria",
         "drinkID": "337e34ba-3d7e-4033-b98a-2ede6ed80ce8",
         "description": "Laphroaig, Brennevin Aquavit, Lemon, House-Made Cardamom Ginger Vanilla Syrup, Angostura Bitters",
@@ -10,7 +10,7 @@
     },
     "-NpHsYVvtj0emdOgk-Mo": {
         "barName": "Stampede Cocktail Club",
-        "barID": "1ea8beea-6dd9-48ea-b279-f40efd36fdf6",
+        "barID": "-Nr7lOhIhP11igBsjqOp",
         "drinkName": "Earthly Tether",
         "drinkID": "345e52c4-edb8-443e-888a-fed00574f7b9",
         "description": "Maitake Mushroom Infused Bourbon, Cynar, House-Made Thyme Tincture, Black Walnut Bitters",
@@ -18,7 +18,7 @@
     },
     "-NpHrmTBbIekPIk0Fd89": {
         "barName": "Wero",
-        "barID": "cb515f4c-1538-4704-b0ab-aa0c586f6b29",
+        "barID": "-NqtP2KgP0su3VWrXHaf",
         "drinkName": "Booryoung",
         "drinkID": "f07f465a-520b-4bb6-b509-3ca83e66f95e",
         "description": "High West Double Rye, Amargo Vallet, Amaro Brovo Kim, Banana, Honey, Angostura",
@@ -26,12 +26,21 @@
     },
     "-NpHs4DwzB9U8ka98ZFb":{
         "barName": "Wero",
-        "barID": "cb515f4c-1538-4704-b0ab-aa0c586f6b29",
+        "barID": "-NqtP2KgP0su3VWrXHaf",
         "drinkName": "Sumnara",
         "drinkID": "7cdfc7c0-bb09-48bb-94cf-2be9860fe30d",
         "description": "Haku Vodka, Jinro Ilpoom Soju, Vanilla, Coconut, Honey, Lemon, Matcha, Fever Tree Mediterranean Ginger Beer",
         "price": 20
     }},
+    "bars": {
+        "-Nr7lOhIhP11igBsjqOp": {
+            "barName": "Stampede Cocktail Club",
+            "barID": "-Nr7lOhIhP11igBsjqOp"
+        },
+        "-NqtP2KgP0su3VWrXHaf": {
+            "barName": "Wero",
+            "barID": "-NqtP2KgP0su3VWrXHaf"
+        }},
 "comments": {
    "-NpI6y26xFs-j2y0M7r9": {
         "userName": "Jessica",
